vfs/cbe: add rekeying
As reykeying is a background job there is not always a client request
pending. Whenever the backend is generating a I/O response send a I/O
signal to the VFS server (I am not sure how it will work out when the
plugin is used in a local VFS), which in return will execute the I/O
signal handler that will execute 'handle_frontend_request()'.

The next step is splitting request handling into bottom and top half
where the bottom half is able to make progress without being triggered
by the top half to clearly indicate when the plugin was triggered by
a client request.