CBE SB control: don't discard non-dispensable snap
If, by now, a user called "discard snapshot" for a snapshot that was the
currently active snapshot or the last secured snapshot, it would have been
simply deleted leaving the CBE device in an inconsistent state. The currently
active and the last secured snapshot must never be deleted.

The problem was that "discard snapshot" used to directly set the "valid" flag
of the found snapshot to "false". Now this was changed to merely set the "keep"
flag to "false". The "keep" flag indicates whether a snapshot is of personal
interest for a user and prevents the snapshot from beeing deleted during the
common garbage collection that is done regularly in the CBE. The garbage
collection is also called in the further course of "discard snapshot".
Therefore, the now non-"keep"-marked snapshot will eventually be also deleted.
But only as soon as it becomes dispensable for the CBE.